Add COMPANY_SERVICE_BASE_URL to config.js
This commit adds the COMPANY_SERVICE_BASE_URL constant to the config.js file in the frontend/src/config directory. The constant is set to 'http://localhost:8000/api/course/company' and will be used for making API requests related to companies. This change enables the frontend to communicate with the company service API.