feat: Send notification on course enrollment and user registration
This commit adds the functionality to send a notification when a user enrolls in a course or registers on the platform. The notification includes an email to the student with relevant information. The notification is triggered using the axios library and the notification service URL specified in the environment variables. This feature enhances the user experience by providing timely updates and confirmation of actions.